+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* New cop `Rails/InverseOf` checks for association arguments that require setting the `inverse_of` option manually. ([@bdewater][])
 * [#4650](https://github.com/bbatsov/rubocop/issues/4650): Add new `Style/StringHashKeys` cop. ([@donjar][])
 * [#1583](https://github.com/bbatsov/rubocop/issues/1583): Add a quiet formatter. ([@drenmi][])
 * Add new `Style/RandomWithOffset` cop. ([@donjar][])
@@ -3018,3 +3019,4 @@
 [@michniewicz]: https://github.com/michniewicz
 [@bgeuken]: https://github.com/bgeuken
 [@mtsmfm]: https://github.com/mtsmfm
+[@bdewater]: https://github.com/bdewater